feat: enable pooling allocation when possible
This turns on pooling allocation if the host supports it.

For now, we use the default pooling settings. These are likely enough, though Spin and `wasmtime serve` differ slightly in their implementations. I think it is fine for now to start with the defaults, and we can adjust if we run into any issues.

If you are interested at all in what this is doing, and what all of the settings are, please read: https://docs.rs/wasmtime/latest/wasmtime/struct.PoolingAllocationConfig.html