refactor: simplify file watcher setup
Now that the cache checks if the file has been modified to drop the cache entry, we don't need to hijack the tag for updating this.

We still keep the watch mode because it needs to watch the directory for files being added or removed, but it doesn't need to keep track if existing files have been updated or not.