refactor: Use rayon instead of tokio::spawn_blocking for CPU intensive tasks
Looking into how spawn_blocking works, they have a very high default for max blocking threads (512).
This is likely because Tokio uses it internally for sync I/O operations that they wrap in async interfaces, so these operations are not CPU intensive.

However, we are trying to shove very CPU intensive tasks into this pool, and if we load a lot of them at once, this could be bad when we run the Kernel in more contrained environments.

This moves it to using Rayon instead for this, where we get a bit easier control over how the CPU thread pool is setup and working.