fix: env filter no longer always set to trace
Right now, if the user doesn't supply a LOG_LEVEL, this env filter isn't valid and falls back to TRACE, which creates a lot of logs.
I simplified the logic a bit here, but we might be able to still do something more sophisticated here.

That being said, I think it is likely that INFO as default is likely fine.