feat: convert AppConfig to builder lite pattern
This moves the fields from AppConfig to be private. This means new
AppConfig fields won't require a breaking change every time. This is
nice because we usually provide a default in order to not break the helm
chart, and this should help avoid confusion from our major version
breaking with each new config item.