feat: allow running without requiring bearer token
if running without a auth url set, we do not expect users to provide
bearer tokens. in case a skill is trying to access resources that
are protected by a PhariaAI token, the clients try to convert the
auth value to a string and will raise a runtime error if that is
not possible.