feat: remove explain functionality in 0.4 wit world
explainability is only supported by the luminous inference
worker. this is being replaced with vllm workers. currently,
there are no plans to implement explainability for vllm, so we
do not want to encourage customers to build on this feature for
future skills. skills written against earlier wit world will keep
working, as long as the customers have a luminous worker running