test: remove assertions against DocumentIndex search endpoint
the DocumentIndex powered the semantic search functionality of the
engine. It is set to be deprecated end of year and does not seem to
be fully funtional currently. This change removes some of the
expectations that break with current DocumentIndex deployments.
Going forward, we'll need to remove the DocumentIndex functionality.
Document (metadata) retrieval by name still seems to be working by now.